Move ModelLoader classes to model.loader package
