+++ src/main/java/com/example/aot/AOT.java
@@ -3,12 +3,11 @@ package com.example.aot;
 import com.example.auxiliary.Timer;
 import com.example.core.model.GGUF;
 import com.example.core.model.tensor.GGMLTensorEntry;
-import com.example.loader.weights.LlamaModelLoader;
+import com.example.model.loader.LlamaModelLoader;
 import com.example.model.Model;
 import com.example.Options;
 import com.example.model.format.LlamaChatFormat;
 import com.example.model.llama.Llama;
-import com.example.loader.weights.ModelLoader;
 import com.example.loader.weights.Weights;
 import com.example.tokenizer.impl.LlamaTokenizer;
